Get bytecode compset working for initial env
This was achieved by throwing in the primitive definitions of

        ,bc_fetch_aux_def
        ,bc_find_loc_aux_def
        ,is_Label_def

and removing the bc_fetch_aux_0_thm and adding the bc_inst theorems to
the type base. This works in the simple case I'm trying to evaluate
bytecode in, but I don't know if it is an improvement or a regression in
general.