performance hacking snoc_cenv_eq_thm Dtype case
The slow part previously was the evaluation of
  ALL_DISTINCT (MAP FST (build_tdefs ... ++ ...))

The change is to use a custom compset instead of EVAL, where this
compset only has lists, strings, pairs, and build_tdefs.

The key, however, seemed to be to remove LIST_TO_SET_THM from the list
compset, because it ruins the nice rewrite for MEM. This should probably
be considered upstream.

This commit also includes some commented out debugging stuff, and a
commented out previous attempt to speed things up by caching
ALL_DISTINCT theorems.