fix broken calls to EVAL in ml_translatorLib.finalise_module_translation
Another case of the_compset containing too much or the wrong stuff,
partly to do with problems with MEM and partly to do with prim_sem_env
(which should not be evaluated according to its definition).

I added two different compsets here, because if I try to combine
everything in one, then one of the usages blows up. Not exactly sure
why.