work on cleaning up repl_funProofScript.sml
move various things to places where they are more likely to belong.  the
repl correctness lemma builds now, but the main repl correctness
theorems are not yet proved (or stated) and there is some problem with
the setup of the initial bytecode state (different places expect
empty_bc_state to have different code).