Mostly fix translator
Commented out

DeclAssumCons_SNOC_Dtype
DeclAssumCons_cons_lookup

because of the change of type of the constructor environments from lists
to finite maps. Correspondingly, DeclAssumCons lost an ALL_DISTINCT
assumption. I don't know why it was there, unless it was an artifact
from the past when duplicate constructors causes the evaluation to type
error.