more thoughts on adding constraints to models
Ran into a (probably minor) problem:
A type definition introduces both a new type and some new constants. If
the type is constrained but the constants are not, there is no guarantee
that the default (non-constrained) semantics of the constants will be
suitable for the new (constrained) value of the type. So there will
probably need to be some condition on the constraints, saying that if
you constrain a type, you also have to constrain any constants whose
types include that type.