bit more on add_constraints_thm
Need something like semantics_extend but going in the other direction:
if you satisfy something in an extended signature then you satisfy it in
a reduced signature. The problem is satisfaction in an extended
signature looks weaker because it quantifies over fewer valuations. But
I think this shouldn't actually matter...