rename various translator/other-examples that depend on word_prelude
the problem with word_prelude is that the translator gets confused about
whether to use the built-in WORD8 refinement invariant for word8, or the
(admittedly poor) WORD refinement invariant defined in word_prelude. the
right way to fix all this is probably to add proper support for word8
operations to CakeML, and just use WORD8 everywhere.