delete out-of-date compiler tests
There may be some interesting small example programs here, but the AST
they are written in is quite out of date and using the parser on tests
written in concrete syntax is probably a better approach for the future.
There is also some broken infrastructure for putting the compiler
through EmitML, but using the translator is a better option for that.