remove old translaton of chars from std_prelude
There is a theorem about CHR on n MOD 256 whose functionality I have not
duplicated in the new setup. I'm not sure if it is necessary to do so...
Indeed, it may be good to look through all that is removed in this
commit and make sure it has been adequately replaced in the previous
commits.