Modify the declarative type system to make the inferencer complete.
Enforce a restriction on declarations that are not syntactic values (and
so can't be generalised). If there are two different types that it could
have, then it has no type. All the metatheory goes through unchanged,
except for weakening, where there is stil 1 tricky cheat dealing with
weakened module environments possibly adding extra types.