initial thoughts on BVL -> bytecode
define a compiler (minus the primitive ops) and sketch part of the
correctness theorem statement. the statement will need to be extended to
have conclusions for all the compilation functions. or maybe the
expression compiler should be reworked to operate on lists of
expressions (like the BVL semantics) to reduce the number of functions...