fix fo_pat_correct
It seems like many proofs that were based on induction on the exp_pat
type will now be broken because they require inductive hypotheses about
doing the semantics on REVERSE of a list of expressions and not just the
list as given. This is the first such proof. Not sure if there's a
better way to proceed...