prove bvl_bc's code loc translation function can be changed
as long as it agrees on the code locs actually appearing in the input.
also, show that the code locs produced by cComp are contained in the
code locs of its input, provided there are no App SOMEs, Calls, or
Labels.

hoping this will allow the main correctness proof (for the REPL
compiler) to union in the function that translates previously defined
code locs.