Good progress on partial function application
Worked through the application case enough to be confident that the
evaluation order stuff works. Had to change the location of closure
arguments to functions in the code table to be the first argument rather
than the last (e.g., Var n for an n argument function, rather than Var 0
always). This is a bit more complicated, but the advantage is that App
(SOME l) x xs can be compiled to Call (SOME l) (xs++[x]), instead of
Let (xs++x) (Call (SOME l) [Var n; Var 0; ...; Var (n-1)] ). The App
NONE case requires a let since the closure must be evaluated last, and
only once.