Progress on the App NONE case.
The clock setup needs to change.

if you have a multi-argument function f

val f = fn <x,y> => ...;

and partially apply it

val p = f 1

and then apply that

val x = p 1

the closLang semantics (and patLang semantics) ticks twice, once for
each arguments. However, the conversion to bvl needs to tick three
times. The tick for f 1 cannot be deferred because that could be the
tick that runs out of fuel. Now p is a partial closure that stores the
argument 1, a pointer to the closure for f, and a code pointer to a stub
function that knows how to set up the call to f. So calling p involves
for a BVL call to the stub, which in turn makes a BVL call to the code
pointer stored in f's closure. So that's two ticks.

One idea is to have the clos->BVL translation guarantee that the BVL
program does exactly twice as many ticks as the closLang program. Then
the above example should do 2 ticks for the call to f, and 2 ticks for
the call to p.