wip attempts to get compile_special and compile_initial verified
not sure that compile_special will still even be needed though - could
just use compile_top (if I recall correctly the reason was something to
do with labels..)?

given that compile_top and compile_prog are verified, it's almost
certainly true that compile_special and compile_initial will be correct
(maybe with minor bugs) but pushing through the proofs will be a bit
tedious.