start updating pmatchExamples
comment out the old version.
new plan is to write the desired PMATCH expression manually (and prove
the equivalence manually, but hopefully a simple tactic will work most
of the time).

this commit: do a new version of raconv_PMATCH following this plan.

however, the translation of this raconv_PMATCH seems to go into an
infinite loop...