compilation of the kernel module to labelled bytecode
in the logic. mostly to see that it works.

could do more (e.g., remove labels, compile further to x86) but it's not
clear that this is the best way to set the kernel module up (i.e., as
the first module after the prim_types_program). it will also eventually
need a signature.