Start fixing the unverified Haskell implementation
It's badly out of date, but I want to use it for generating good
location-aware error messages to help in porting to CakeML. I also want
a CakeML to SML and OCaml translator to enable us to benchmark. NB, the
interpreter isn't so important, for this but the lexer, parser, and type
inferencer are.