Trying to improve the closLang semantic relation
We want flexibility for the compiler to change the number of clock steps
required, but we have to be careful that relation implies divergence
preservation. The relation here is trying to use an unbounded, monotone
function, but the quantification structure doesn't work out. I'll have
to try something more complex.