Refactor conversion of binary operators.
Converted ASTs are the same for the moment, but I intend to correctly
convert them to applications to tuples in a forthcoming branch and this
abstraction (a mk_binop function) should make this easier.

Also use "return" rather than "SOME" in a bunch of places to make
possible switch to a different monad a bit easier.