Set up (as yet unused) monad for handling fixities
Enabling this will require fairly drastic (but simplifying, so there's
that) surgery on the grammar, with effects on the soundness and
completeness proofs of course.

Also put in a place-holder for storing constructor arities, for fixing
the currently broken handling of those (writing SOME(1,2) breaks).
Knowing what constructors were around (those with arities recorded)
would also allow for SML-like categorisation of identifiers (rather than
relying on the capitalisation).