Custom versions of choicel + pegf for translation's sake
The more polymorphic versions in the core distribution use ARB and
can't (it seems) be translated. I had earlier thought that a
type-specific version of the definition would work, but though the
behaviour of versions with [] in the place of ARB would be the same, the
peg symbols are still different, and so the equation (which is on the
syntax, not the behaviour) is not true.

bootstrap/translation/reply/replMLScript now gets past the parsing
constants.