initial work on monadic OpenTheory article reader
progress on #69

This commit defines a monadic function (within the HOL Kernel's
state-exception monad) that takes a line from an OpenTheory article and
the article reader's state, and returns the new article-reader state (or
fails). Since the HOL monad cannot easily be extended, the article state
is passed around functionally.

The main thing missing from the function is proper treatment of
situations where OpenTheory specifies conversion to a particular
representative of a set of theorems (up to alpha-equivalence, and
allowing hypotheses to be added): currently, we just use the original
theorem instead.

The main functionality missing is a driver that hooks this line-reading
function up to a real input stream and iterates it, and something to
print the output list of theorems at the end.

The semantics of the reader is less general than OpenTheory in that we
only allow articles whose axioms are those in the initial context of
HOL, i.e., the reader does not permit real new axioms, or those that
would be encountered in little theories. The idea is for all theory
composition to be done externally, and the final composed article fed to
this reader.

Derived rules SYM, TRANS, and PROVE_HYP need to be either added to the
kernel (I elsewhere have proofs that they are sound for this purpose) or
implemented as derived rules.

This is untested and nothing is proved about it yet (except that it type
checks).