some derived rules (sym, trans, proveHyp)
taken from hol-reflection repo.

I just realised trans is already in the kernel, but this shows that it
doesn't have to be proved sound against the set-theory semantics
directly. I put a TODO suggesting taking trans out of the inference
system (but it should still be in the implementation of the kernel, for
efficiency).