Start refactoring the semantics some
Tried to make the functional big step as readable as possible.
Curried everything, changed some argument ordering, and renamed some
things. Also put environments into a record instead of a tuple, and made
the state record have everything in it (including stuff that expression
evaluation does not look at). The only exception is do_app which just
takes the refs and io.

Some semanticPrimitive functions had to change which had knock-on
effects in the small and relational big-step semantics.

The proofs don't build.