Prove log reln implies preservation of obs. sem
Fix as per Magnus's recent post, which asserts that we're only
interested in situations where the expression-state pair doesn't
type-error (on any clock, or any IO trace).  This is captured by the

   ~sem e1 s1 Fail

hypothesis.

The theorem is slightly stronger because it allows for a state_rel
between the starting states, rather than requiring exactly the same
state.

Also delete a lot of unrelated cruft