clos$Fn has num list option to record its free vs
This is first step towards removing the need for the restrict_env
component of the state because the syntax now records whether or not the
clos_free optimisation has been performed. The semantics respects this:
if the field is NONE, the environment is the whole of the global
environment, otherwise it looks up in the global environment with the
provided list when it creates the closure value.

Something analogous still to be done for Letrec.

It would be nice if there was an easy to run selftest that somehow ran
some test-cases to suggest whether or not the change appeared valid.