more lem issues; workaround: remove polymorphism in store_ffi
the previous problem was that lem fails to do anything useful with val
declarations (which are in effect type annotations) in the hol backend.

the problem addressed here is that lem assumes type arguments for
abbreviations are positional whereas in hol they are alphabetical.

both are known lem issues that will probably never get fixed.