Prove "backend_correct" for MIPS.
There is now just one "backend_correct" predicate. The simple version (which assumed there are no processor exceptions) is replaced with one that considers side-effect from a (suitably constrained) "env" function, which is interleaved with an ISA next-state function.

A simplification is made to handle the MIPS branch-delay slot (see the definition of "mips_next"). As such, "env" is never applied immediately after a branch. On MIPS an exception can be taken when there is an instruction in the delay slot (i.e. after a branch) and the exception handler is then responsible for correctly replaying the branch (i.e. returning to EPC, which is PC - 4). The "backend_correct mips_target" proof does not cover this behaviour, so the correctness assertion is based on the implicit assumption that this case is handled correctly.