prove semantics_deterministic, but had to add to can_type_prog
This is determinism for the top-level semantics function. Since it is
already a function the only source of non-determinism is in the set of
behaviours if it returns an Execute result. We here prove that that set
is always a singleton. However, we require an assumption of the state
(the type soundness invariants assumption), which hopefully we can prove
and discharge for the initial state.

The additions to can_type_prog might be removable somehow, but I'm not
sure; I suspect the reason for the addition is because the definitions
are all a bit out of sync (e.g., prog_diverges vs evaluate_whole_prog).