Try defining modSem top-level semantics as a function
And try proving a top-level correctness theorem. didn't get very far.
It's quite a mess, because we need to know modLang's semantics is
deterministic and monotonic etc.

It looks like it would be worthwhile to convert modSem to a functional
big-step semantics before continuing.

Also, it's not clear whether it's worthwhile having the compiler return
its final state/config. What are we going to prove about that? What do
we need to prove about it for when we add eval?