Add backendTheory, defining the whole backend.
Replaces the _to_target theories. Adds a top-level backend config
record. Compilers for each intermediate language are easily defined from
here (but not currently included).

A few changes made to the interfaces of some parts of the compiler along
the way. The definition of the compiler may still be wrong; indeed, I
think I noticed (and fixed) some problems regarding how the config
parameters are used. It's possible that some config parameters should
instead be hard-coded in the compiler interface (i.e., are not really
configurable).