make bvpSem$get_var(s) take just the locals, not the whole state
also fix an error (order of declarations) in bvp_to_word, to get it to
build.

the same issue (taking too big an argument) afflicts the get_vars in
wordSem, but this is left as is for now. arguably get_vars shouldn't
exist at all (it's the same as lookup).