Fix definition of (clos_)remove in Letrec case
Error fixed was to imagine that one could drop bindings for those
functions that don't appear in the body, but these functions may of
course be called from other functions that *are* called in the body.

So, follow original definition, which drops all bindings when none occur
in the body. Strictly speaking, we might imagine doing an analysis based
on the topological sort of the bindings and drop uncalled
sub-components.

Prove this version correct.

Closes #93