define primitive compiler config, and prove some cheats in backendProof
the precondition on backendProof is that the supplied config agrees with
the primitive compiler config in certain places (but can differ
otherwise, to allow custom configs).

there is a cheat that looks false to do with assumptions on the
mod_to_conProof - the theorem or the compiler is wrong in its
treatment of the primitive option type. the theorem is probably too
specific.