delete list_max and max2
use list_max from miscTheory (which should probably be moved to HOL at
some point). The version without the accumulator is probably slightly
less efficient, but a version with accumulator can easily be substituted
in for EVAL and translation if it proves to be important.

max2 is just MAX which is already in HOL