only one cheat left in bvp-to-word, no cheats in word-to-word
The remaining bvp-to-word cheat is assign_thm, i.e. implementing
BVP's ops in wordLang. Note that my latest change forbids the
op implementations from requiring extra ticks. Yong Kiam and I
have drawn up a plan for how this can be done even though some
of the op implementations will require loops.

Yong Kiam and I plan to add a new construct to wordLang. The
new construct, MustTerminate n prog, allows prog to run with
clock set to n, and restores the clock when MustTerminate
returns. Note that a TimeOut within evaluation of prog causes
an error. In other words, we allow prog to require up to n
ticks, but make it seem as if no ticks are required.