change semantics of W8toInt, to match SML's Word.toIntX
i.e., make it the signed version. I'm not sure if both primitives are
necessary: probably one can be implemented in terms of the other. If so,
I am picking the signed version for the primitive since then it is
inverted by W8fromInt. (If not we can add the other primitive easily.)

compiler and proofs not updated yet.