Replace term_rewrite with a better implementation
The problem with the version with ASSUME is that PURE_REWRITE_CONV will
not instantiate type variables in the assumptions. The new version is
equivalent to the one in the comments (that did the match and subst
manually) but shorter because it uses PURE_REWRITE_CONV to do the work.