Update preamble after moving some things into HOL
In particular:
  split_pair_tac         -> pairarg_tac
  split_applied_pair_tac -> split_uncurry_arg_tac
  split_pair_case_tac    -> split_pair_case0_tac
  split_pair_case_tac++  -> split_pair_case_tac

The behaviour is slightly different because of things like searching in
the goal first rather than the assumptions, and not forcing so many
variables to be primed. Therefore some proofs needed minor fixes.

The split_pair_case_tac++ thing above means doing a search across the
whole goal for split_pair_case0_tac opportunities; but the
implementation in HOL seems currently to be broken, so the places where
it should be working are fixed with worse proof but marked with TODO.