Make bvp_live delete dead code; run bvp_simp after bvp_live
The bvp_live transformation can only delete pure expressions.
Unfortunately, Cons and arithmetic (Add, Sub, etc.) cannot be
deleted because they touch the space component of the state,
i.e. they are stateful.

In the future we might consider adding a switch to the BVP
semantics which allows turning off the space component.
The bvp_space optimisation (which is run after bvp_live)
could turn on the space state component.