Define WORD ref inv, prove translation thms
The WORD refinement invariant maps word8 to Word8 in CakeML,
and word64 to Word64. All smaller words are mapped into
appropriate versions of either word8 or word64. Words larger
than 64 bits are not supported at present.

The theory has translation theorems for n2w, w2n, and, or,
xor, and all of the shift operations. (Comparison operations
are currently missing.)

The current translation won't be able to support translation
of the 'a word type. However, it will be able to support all
instances of 'a word that fit in a word64. (I could have used
a different scheme that would allow translation of 'a word, but
any such scheme would not have mapped word8 to Word8 in CakeML
etc.) In practice this means that various compiler definitions
need to be (automatically) type instantiated before they are
handed to the translator.