Prove results about known approximations
The system is very akin to a type-system with sub-typing. The way the
globals map gets updated is still annoying: the algorithm is
incomplete (doesn't handle long chains of updates through globals) and
correctness is probably reliant on the fact that there is never any more
than one SetGlobal call for any one global.