Add more test-cases for 'known' in a comment
These (labelled "2B" and "2C") demonstrate why I'm coming to loathe
'known'. Test 2B shows that if the set/get pairs come in the right
syntactic order, then 'known' does great things, with the flow of
setting and getting correctly populating the g-map.

However, the attempt to support the backwards-looking idiom in "TEST 2",
makes 'known' do some very strange things. Nor is it clear that it wins
very much. The example in TEST 2C will never see 'known' infer a good
approximation for global-62 because the first time around it infers
"Other", and Other is at the top of the merge lattice, meaning that all
subsequent updates to the g-map at key=62 will keep the value as Other.

Further comment to come in an e-mail.