Prove properties of known with restricted SetGlobals
In particular, show that if the g-map doesn't have anything to say about
the SetGlobals that occur in a list of expressions, and there aren't any
more than 1 SetGlobal for any given index, then 'known' will grow in a
predictable way, extending the g-map cleanly.

This result (known_better_definedg) is all very well, but its
preconditions are immediately violated by the two successive calls to
'known' in the clos_knonwn$compile function precisely because the g-map
calculated by the first is passed as an argument to the second.