Complete App case of known's correctness theorem
This case requires a slight change to statement of result. It's possible
that the optimisation here will make the program more defined because an
App NONE can fail with a type error if there are too many
arguments (more than the max_app constant), but App (SOME loc) can't.
The point of the known optimisation is too introduce the latter forms,
so programs that used to have type errors will now execute safely.