Make minor cleanups ahead of need to recast proof
Approach with equality is never going to cut it: had forgotten that
known optimises in bodies of abstractions, meaning that returned values
can't be equal when evaluating Fn and Letrec. D'oh!

The existing logical relation isn't parameterised by states satisfying a
particular predicate, so a custom, syntactic relation is probably the
way to go.