Prove part of Word64 ops in bvp_to_word
word64 proved. word32 partially proved (cheated), although Add and Sub
are not implemented yet.

The cheat would probably be easier if I used more (unnecessary)
temporary variables in the implementation. But I'm not sure if I can
count on later parts of the compiler to re-use the same temporary again
(and if I should be wary of giving them extra work).