Refine syntactic relation for known-correctness
Still problematic, mostly because the syntactic connection between two
expressions (the bodies of closures) depends on
parameters (environmental approximation and approximation information
for globals), and it seems as if these parameters have to appear as
parameters to the relation; then they vary in the inductive hypotheses.

Handling the way the environment approximation changes in evaluating Let
is the current stumbling block.