Tweak condition on len_size, prove memory_rel_ByteArray_IMP
The condition in heap_in_memory_store_def appears to only be required
for the byte array lemma proved here (changing it down to 5 from 7 does
not affect other bvp_to_word proofs, although this may be because it is
hidden by conf_ok).

I am unsure why only 7 (and not 8 for :'a = :64) is necessary.