Prove more cases, but approach is still doomed
Magnus suggested insisting that the code tables be empty, and this does
make the Call case go through, and this will certainly make the eventual
proof easier, but the value relation is still wrong (evaluate_app
reveals the problem). I'm back to being convinced that the various
relations need to be parameterised by the env-approximation and globals
approximation information (à la what things looked like in de4c64918b5).