Prove UpdateByte with two serious cheats on set_byte
The strategy was to extend memory_rel_ByteArray_IMP with a clause for
updating a byte (in addition to the one for reading a byte).

The remaining cheats include one that looks reasonable (but maybe hard):

  set_byte (n2w i) b (bytes_to_word (2 ** k) 0w ls t be) be =
    bytes_to_word (2 ** k) 0w (LUPDATE b i ls) t be

and one that looks unreasonable:

  set_byte k b w1 be = set_byte k b w2 be ∧ (w1 = w2 ⇒ v1 = v2) ⇒
    set_byte k b1 v1 be = set_byte k b1 v2 be

but this second one is unnecessary if one can prove:

 write_bytes bs w1 be = write_bytes bs w2 be ∧
 LENGTH w1 = LENGTH w2 ∧ LENGTH bs' = LENGTH bs ⇒
   write_bytes bs' w1 be = write_bytes bs' w2 be

which seems more plausible.