Tidy up known_correct proof to get a cheated build
I think this is the best state so far for the g-parameterised approach.
It fails because the expression extracted from the left-hand-side of a
Full_app evaluate_app case is only known to be related to the second
expression via a gmap g0 that is less-defined than the gmap g that is
used to relate the environments.

I.e., we have

   known [e1] as g0 = ([(e2,apx)], g')
   LIST_REL (kvrel g) env1 env2
   better_definedg g0 g

and the e1 and e2 are executed in environments env1 and env2
respectively. We have (and have earlier relied on) the fact that

  LIST_REL (kvrel g0) env1 env2 ⇒ LIST_REL (kvrel g) env1 env2

but the reverse doesn't hold, and so our IH isn't usable.

Something important is going to have to change.