Work on verifying bvp_to_word FFI
Currently, there is a cheat that looks false and has something to do
with the names set being passed in to assign (and then on to the FFI
instruction), which I do not understand very well.

Reading the byte array seems done, but the writing part still needs
verifying, and it is hard to get to it while the names set thing is
still in the way.