Prove a correctness result for clos_known$compile
An explanation of what the various conditions entail.

In the precondition:
- esgc_free e0 ⇔ no set-globals inside closures within e0
- every_Fn_vs_NONE ⇔
    no precomputation of free variables for Fn and Letrec
- ksrel LN s01 s02 ⇔
    - s01 and s02 (states) must not have code tables
    - s01 and s02 must have identical clock and ffi components
    - their globals must be value-related (but see sga condition (next))
    - their refs must be value-related
- state_globals_approx s01 LN ⇔
    s01 can't have any defined globals
- BAG_ALL_DISTINCT (set_globals e0) ⇔
    e0 can only call SetGlobal n for any n in at most one position
- ssgc_free s01 ⇔ no SetGlobals inside closures anywhere in s01

In the conclusion

  krrel g (res1,s1) (res2,s2)

relates identical values and aborts in res{1,2} to each other, except
that if res1 is a type-error, then res2 can be anything, and if res1 is
a closure value, then the body of that closure may be related by
kerel (which essentially says that the body of the second is the result
of a known transformation of the first).  If such closures are applied,
to related arguments then they produce related results (second conjunct
of known_correct wrt evaluate_app).

The preconditions mentioning LN might be changed if the computed global
approximation maps used inside compile were allowed to leak out.