Start looking at new definition of calls with ghost state
The new version keeps track of which locs it has _not_ added to the code
table, because the relevant function body contained free variables. The
hope is that correctness for this version will be easier because we can
refer to the discarded locs. Although, maybe it would also be possible
to consider the difference between the locs in the expression and the
ones picked up by the original calls function...?