Add another argument to clos_callProof$v_rel
The new argument carries the locs that should not be allowed to go into
the call table (because they correspond to real closures). If something
is in the exclusion set, it cannot be in the g set maintained by the
calls function.

Tracking this information can be done in a neatly monotonic way, and
allows to make minor progress on the App case: now I know the loc that
is trying to be Called actually exists in the table. Proving that it
points to an appropriate thing is the next hurdle...