Make closLang's Call tick more like BVL's
This allows multiple ticks to be made between evaluating the arguments
of a call and the call itself without introducing an additional Let,
which is useful if an App is to be turned into a Call.

Update proofs, including removing some simple cheats in clos_callProof
that required this. (It has not been used to full extent in
clos_callProof yet.)