Prove App in clos_callProof after adding existential clock
Update the other bits of proof to deal with the additional clock ticks
in the new inductive hypothesis (since most of it was cheated anyway,
this wasn't too hard).

Strengthen dest_closure_v_rel_lookup, the main lemma for the App case.

Change the number of extra ticks on the Call expression introduced by
clos_call... this was confusing (it seems like the wrong number) but it
makes the proof work...