Tweak v_rel and pull out closure rels
The induction for calls_correct is wrong and looks possibly non-trivial
to fix. Pulled out the closure rels in the hope of exposing the
existentially quantified g0 (the original argument passed to call when
the closure-creation code was produced) which may need to show up in the
evaluate_app inductive hypothesis.