Remove last cheat in clos_to_bvlProof
A revised obligation to show that initial_state satisfies the
state-relation will show up somewhere (backendProof I suppose). This
will be where the fact that known wants no code and no globals will show
up. In order to show that there are no set-global closures in the state,
it will also be necessary to have that the refs are empty.  (All these
are true of bvlSem's initial state, so it should be straightforward to
push that fact back through full_state_rel.)