Prove one of the evaluate_app cases
In particular: when the app is results in error and was not optimised by
the calls phase. This is one good test of the inductive hypotheses.  A
more tricky one will be the case where the app was optimised by the
calls compiler. If that works, I expect most other things to be just
tedious or requring minor tweaks.