Try a new value (and env) relation in clos_callProof
This breaks many of the cases Magnus proved (but should not be too hard
to fix). The App and Fn cases have been reproved; the hope is that the
new relation makes the evaluate_app case possible, but it's not obvious
yet that it's good enough.

Also deleted some old commented out proof, and added in comments some
example programs that motivate the need for a more flexible value
relation than previously.